+++ ChangeLog.md
@@ -5,6 +5,7 @@ where provided, and the git commit history.
 
 * 2017/08/?? - 5.02 - PENDING
     * Removing vuid checking for XML import 
+    * 508 - changing tabbing color to red when inside of panel heading + Button Edit Concept focuses on first input when opened
 
 * 2017/08/25 - 5.01
     * Updates to parallel translations during first Komet request